refactor ui modules
